requirements.txt
flask
flask-cors
pandas
mysql-connector-python
pymongo